requirements.txt
openai
requests
RPi.GPIO 
